Streamlit/requirements.txt
pandas
streamlit
#streamlit-files-connection
gcsfs
plotly
db-dtypes
nltk
google-cloud-bigquery
google.cloud
google.auth